test/curator-lambda/code/requirements.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
elasticsearch==6.0.0
elasticsearch-curator==5.4.1
idna==2.6
PyYAML==3.12
requests==2.18.4
requests-aws4auth==0.9
urllib3==1.22
voluptuous==0.10.5
